pyproject.toml
[tool.poetry]
name = "archivist-bot"
version = "0.1.0"
description = ""
authors = ["Reliable Magician <petr.b.lavrov@gmail.com>"]
readme = "README.md"
packages = [{ include = "archivist_bot" }]

[tool.poetry.dependencies]
python = "^3.10"
gpt-kit = { git = "https://github.com/calmmage/gpt_kit.git", branch = "main" }
#calmlib = { git = "https://github.com/calmmage/calmlib.git", branch = "main" }
python-telegram-bot = "^20.3"
pydantic = "^1.10.8"
loguru = "^0.7.0"
notion = "^0.0.28"
notion-client = "^2.0.0"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
